backend/pyproject.toml
[project]
name = "crsp-backend"
version = "0.1.0"
description = "CRSP - Course Registration and Scheduling Platform (FastAPI backend)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136",
    "uvicorn[standard]>=0.46",
    "pydantic>=2.13",
    "pydantic-settings>=2.14",
    "sqlalchemy>=2.0.49",
    "alembic>=1.18",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2",
    "httpx>=0.28",
    "email-validator>=2.3",
    "python-multipart>=0.0.27",
    "pyjwt>=2.12.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4,<5",
    "selectolax>=0.3.27",
    "arq>=0.26",
    "redis>=5.2",
    "cryptography>=43.0",
    "structlog>=24.4",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-instrumentation-logging>=0.48b0",
    "celery>=5.4",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
    "pre-commit>=4",
    "respx>=0.21",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"
